refactor: remove local model support and make project Gemini-exclusive
