+++ .env.example
@@ -7,12 +7,8 @@ SEARXNG_URL=
 TAVILY_API_KEY=tvly-xxxxx      # Get your key at https://tavily.com
 PERPLEXITY_API_KEY=pplx-xxxxx  # Get your key at https://www.perplexity.ai
 
-# LLM Configuration
-LLM_PROVIDER=lmstudio          # Options: ollama, lmstudio, gemini
-LOCAL_LLM=qwen_qwq-32b         # Model name in LMStudio/Ollama (gemini-2.0-flash)
-LMSTUDIO_BASE_URL=http://localhost:1234/v1  # LMStudio OpenAI-compatible API URL
-OLLAMA_BASE_URL=http://localhost:11434 # the endpoint of the Ollama service, defaults to http://localhost:11434 if not set
-GOOGLE_API_KEY=AIza-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx # Gemini API Key
+# LLM Configuration (Gemini API)
+GOOGLE_API_KEY=AIza-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx # Get your key at https://aistudio.google.com/app/apikey
 
 MAX_WEB_RESEARCH_LOOPS=3
 FETCH_FULL_PAGE=True
